LayoutTests/platform/gtk/accessibility/img-fallsback-to-title-expected.txt

     test
test
This tests that images will fallback to using the title attribute if no other descriptive text is present.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Image1:
	AXTitle: test1
	AXDescription:
Image2:
	AXTitle: alt
	AXDescription: test2
Image3:
	AXTitle: test3
	AXDescription:
Image4:
	AXTitle: alt
	AXDescription: test4
PASS imagesGroup.childAtIndex(0).childrenCount is 2
PASS successfullyParsed is true

TEST COMPLETE
